Add publication-quality charts and visualization tools
Generated 6 professional charts in PNG (300 DPI) and PDF formats:
- Chart 1: Boot time comparison (Fort Firewall -14.62% improvement!)
- Chart 2: RAM usage (TotalAV +464 MB overhead)
- Chart 3: App launch (TotalAV +50% worst impact)
- Chart 4: Network performance (dual SMB/FTP comparison)
- Chart 5: Comprehensive overhead heatmap (all metrics at-a-glance)
- Chart 6: Process count (minimal variation)

Features:
- Colorblind-friendly palette
- Academic serif fonts (Times New Roman)
- Error bars showing measurement uncertainty
- Percentage change annotations
- 300 DPI publication quality
- Both raster (PNG) and vector (PDF) formats

Tools:
- Generate-Charts.py with matplotlib/seaborn
- Fixed Analyze-Data.py JSON serialization
- Chart descriptions document for LaTeX integration

+++ scripts/Analyze-Data.py
@@ -37,13 +37,13 @@ def remove_outliers_iqr(data, column):
 def calculate_statistics(data, column):
     """Calculate comprehensive statistics for a column"""
     return {
-        'mean': data[column].mean(),
-        'median': data[column].median(),
-        'std': data[column].std(),
-        'min': data[column].min(),
-        'max': data[column].max(),
-        'count': len(data),
-        'cv': (data[column].std() / data[column].mean() * 100) if data[column].mean() != 0 else 0  # Coefficient of variation
+        'mean': float(data[column].mean()),
+        'median': float(data[column].median()),
+        'std': float(data[column].std()),
+        'min': float(data[column].min()),
+        'max': float(data[column].max()),
+        'count': int(len(data)),
+        'cv': float((data[column].std() / data[column].mean() * 100) if data[column].mean() != 0 else 0)  # Coefficient of variation
     }
 
 def analyze_boot_time():
